renamed some c-related files
